requirements.txt
flask
gunicorn
flask
gunicorn
numpy
pandas
scikit-learn
PyPDF2
ollama
request
